+++ jdaviz/core/loaders/parsers/fits.py
@@ -11,20 +11,15 @@ __all__ = ['FITSParser']
 @loader_parser_registry('fits')
 class FITSParser(BaseParser):
 
-    @property
-    def is_valid(self):
-        if self._app.config not in ('deconfigged', 'specviz2d',
-                                    'lcviz', 'imviz', 'cubeviz',
-                                    'rampviz'):
+    def _check_is_valid(self):
+        accepted_configs = ('deconfigged', 'specviz2d', 'lcviz', 'imviz', 'cubeviz', 'rampviz')
+        if self._app.config not in accepted_configs:
             # NOTE: temporary during deconfig process
-            return False
+            return (f"fits format is only supported in the {', '.join(accepted_configs)} "
+                    f"configurations.")
 
-        try:
-            self.output
-        except Exception:
-            return False
-
-        return True
+        _ = self.output
+        return ''
 
     @cached_property
     def output(self):