+++ jdaviz/core/loaders/parsers/specutils.py
@@ -19,16 +19,14 @@ __all__ = ['SpecutilsSpectrumParser',
 class SpecutilsSpectrumParser(BaseParser):
     SpecutilsCls = Spectrum
 
-    @property
-    def is_valid(self):
-        if self._app.config not in ('deconfigged', 'specviz', 'specviz2d', 'cubeviz'):
+    def _check_is_valid(self):
+        accepted_configs = ('deconfigged', 'specviz', 'specviz2d', 'cubeviz')
+        if self._app.config not in accepted_configs:
             # NOTE: temporary during deconfig process
-            return False
-        try:
-            self.output
-        except Exception:
-            return False
-        return True
+            return f"specutils.Spectrum format is only supported in {', '.join(accepted_configs)}."
+
+        _ = self.output
+        return ''
 
     @cached_property
     def output(self):
@@ -37,11 +35,12 @@ class SpecutilsSpectrumParser(BaseParser):
 
 @loader_parser_registry('specutils.Spectrum(array)')
 class SpecutilsSpectrumArrayParser(SpecutilsSpectrumParser):
-    @property
-    def is_valid(self):
-        return (isinstance(self.input, np.ndarray)
-                and self.input.ndim in (1, 2, 3)
-                and super().is_valid)
+    def _check_is_valid(self):
+        if not (isinstance(self.input, np.ndarray)
+                and self.input.ndim in (1, 2, 3)):
+            return 'Input must be a numpy array with 1, 2, or 3 dimensions.'
+
+        return super()._check_is_valid()
 
     @cached_property
     def output(self):
@@ -62,12 +61,16 @@ class SpecutilsSpectrumArrayParser(SpecutilsSpectrumParser):
 class SpecutilsSpectrumListParser(SpecutilsSpectrumParser):
     SpecutilsCls = SpectrumList
 
-    @property
-    def is_valid(self):
+    def _check_is_valid(self):
         if self._app.config not in ('deconfigged', 'specviz'):
             # NOTE: temporary during deconfig process
-            return False
-        return super().is_valid and len(self.output) > 1
+            return 'specutils.SpectrumList format is only supported in deconfigged, specviz.'
+        result = super()._check_is_valid()
+        if result:
+            return result
+        if len(self.output) <= 1:
+            return 'SpectrumList must contain more than one spectrum.'
+        return ''
 
     @cached_property
     def output(self):